move a bunch of stuff from Func.Api to Core
