+++ match/src/Piipan.Match/Piipan.Match.Api/Models/IRequestPerson.cs
@@ -1,0 +1,7 @@
+namespace Piipan.Match.Api.Models
+{
+    public interface IRequestPerson
+    {
+        string LdsHash { get; set; }
+    }
+}
