+++ match/src/Piipan.Match/Piipan.Match.Core/Builders/ActiveMatchRecordBuilder.cs
@@ -1,7 +1,8 @@
 using System.Text.Json;
 using Piipan.Match.Api;
+using Piipan.Match.Api.Models;
 using Piipan.Match.Core.Models;
-using Piipan.Match.Orchestrator;
+using Piipan.Participants.Api.Models;
 
 namespace Piipan.Match.Core.Builders
 {
@@ -37,7 +38,7 @@ namespace Piipan.Match.Core.Builders
         /// <param name="input">The RequestPerson object received as input to the active match API request.</param>
         /// <param name="match">The ParticipantRecord object received as output from active match API response.</param>
         /// <returns>`this` to allow for method chanining.</returns>
-        public IActiveMatchRecordBuilder SetMatch(RequestPerson input, ParticipantRecord match)
+        public IActiveMatchRecordBuilder SetMatch(IRequestPerson input, IParticipant match)
         {
             this._record.Input = JsonSerializer.Serialize(input);
             this._record.Data = JsonSerializer.Serialize(match);
