+++ match/src/Piipan.Match/Piipan.Match.Core/Builders/IActiveMatchRecordBuilder.cs
@@ -1,11 +1,12 @@
 using Piipan.Match.Api;
-using Piipan.Match.Orchestrator;
+using Piipan.Match.Api.Models;
+using Piipan.Participants.Api.Models;
 
 namespace Piipan.Match.Core.Builders
 {
     public interface IActiveMatchRecordBuilder
     {
-        IActiveMatchRecordBuilder SetMatch(RequestPerson input, ParticipantRecord match);
+        IActiveMatchRecordBuilder SetMatch(IRequestPerson input, IParticipant match);
         IActiveMatchRecordBuilder SetStates(string stateA, string stateB);
 
         IMatchRecord GetRecord();
