+++ match/src/Piipan.Match/Piipan.Match.Core/Piipan.Match.Core.csproj
@@ -8,9 +8,9 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Piipan.Match.Api\Piipan.Match.Api.csproj" />
+    <ProjectReference Include="..\..\Piipan.Match.Shared\Piipan.Match.Shared.csproj" />
     <ProjectReference Include="..\..\..\..\shared\src\Piipan.Shared\Piipan.Shared.csproj" />
     <ProjectReference Include="..\..\..\..\participants\src\Piipan.Participants\Piipan.Participants.Api\Piipan.Participants.Api.csproj" />
-    <ProjectReference Include="..\..\Piipan.Match.Orchestrator\Piipan.Match.Orchestrator.csproj" />
   </ItemGroup>
 
   <ItemGroup>
@@ -18,6 +18,8 @@
     <PackageReference Include="Microsoft.Extensions.Logging" Version="3.1.19" />
     <PackageReference Include="Nanoid" Version="2.1.0" />
     <PackageReference Include="Npgsql" Version="5.0.10" />
+    <PackageReference Include="Newtonsoft.Json.Schema" Version="3.0.14" />
+    <PackageReference Include="FluentValidation" Version="10.3.3" />
   </ItemGroup>
 
 </Project>
