+++ match/src/Piipan.Match/Piipan.Match.Func.Api/Api.cs
@@ -8,10 +8,10 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Azure.WebJobs;
 using Microsoft.Azure.WebJobs.Extensions.Http;
 using Microsoft.Extensions.Logging;
+using Piipan.Match.Core.Models;
+using Piipan.Match.Core.Parsers;
+using Piipan.Match.Core.Resolvers;
 using Piipan.Match.Func.Api.DataTypeHandlers;
-using Piipan.Match.Func.Api.Models;
-using Piipan.Match.Func.Api.Parsers;
-using Piipan.Match.Func.Api.Resolvers;
 using Piipan.Match.Shared;
 
 namespace Piipan.Match.Func.Api
