+++ match/src/Piipan.Match/Piipan.Match.Func.Api/Startup.cs
@@ -2,10 +2,10 @@ using System.Data.Common;
 using Microsoft.Azure.Functions.Extensions.DependencyInjection;
 using Microsoft.Azure.Services.AppAuthentication;
 using Microsoft.Extensions.DependencyInjection;
-using Piipan.Match.Func.Api.Models;
-using Piipan.Match.Func.Api.Parsers;
-using Piipan.Match.Func.Api.Resolvers;
-using Piipan.Match.Func.Api.Validators;
+using Piipan.Match.Core.Models;
+using Piipan.Match.Core.Parsers;
+using Piipan.Match.Core.Resolvers;
+using Piipan.Match.Core.Validators;
 using Piipan.Participants.Core.Extensions;
 using Piipan.Shared;
 using Piipan.Shared.Authentication;
