+++ match/tests/Piipan.Match.Core.Tests/Validators/RequestPersonPiiValidatorTests.cs
@@ -1,10 +1,11 @@
 using System;
 using System.Linq;
-using Piipan.Match.Func.Api.Models;
+using Piipan.Match.Core.Models;
+using Piipan.Match.Core.Validators;
 using FluentValidation.TestHelper;
 using Xunit;
 
-namespace Piipan.Match.Func.Api.Validators.Tests
+namespace Piipan.Match.Core.Tests.Validators
 {
     public class RequestPersonPiiValidatorTests
     {
