+++ match/tests/Piipan.Match.Func.Api.Tests/StartupTests.cs
@@ -1,8 +1,8 @@
 using Microsoft.Azure.Functions.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection;
-using Piipan.Match.Func.Api.Models;
-using Piipan.Match.Func.Api.Parsers;
-using Piipan.Match.Func.Api.Resolvers;
+using Piipan.Match.Core.Models;
+using Piipan.Match.Core.Parsers;
+using Piipan.Match.Core.Resolvers;
 using Piipan.Shared;
 using Piipan.Shared.Authentication;
 using FluentValidation;
